feat: multichain faucet
