+++ frontend/src/components/Home.scss
@@ -1,0 +1,114 @@
+.Home {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  section {
+    max-width: 640px;
+    color: white;
+    background-color: rgba(38, 38, 38, 0.98);
+    border-radius: 2px;
+    margin: 1em;
+    border: solid 1px rgb(40, 40, 40);
+    box-shadow: 0 6px 6px 0 rgba(0, 0, 0, 0.5);
+    box-sizing: border-box;
+
+    h2 {
+      margin: 0 0 20px 0;
+      font-weight: 500;
+      font-size: 26px;
+    }
+
+    article {
+      line-height: 1.4;
+    }
+
+    .InputContainer {
+      ::placeholder {
+        /* Chrome, Firefox, Opera, Safari 10.1+ */
+        color: rgba(14, 60, 165, 0.5) !important;
+      }
+
+      ::-ms-input-placeholder {
+        /* Microsoft Edge */
+        color: rgba(14, 60, 165, 0.5) !important;
+      }
+
+      .fieldError {
+        margin-top: 0.6em;
+        color: #ff5561;
+        font-size: 14px;
+      }
+
+      .input {
+        margin-top: 1em;
+
+        input {
+          appearance: none;
+          background: rgba(255, 255, 255, 0.94);
+          border: solid 1px rgba(255, 255, 255, 0.94);
+          border-radius: 4px;
+          height: 46px;
+          padding: 0 0.6em;
+          width: 100%;
+          outline: 0;
+          box-sizing: border-box;
+          font-size: 16px;
+
+          &:focus {
+            border: solid 1px #859dca;
+          }
+        }
+      }
+    }
+
+    button {
+      margin-top: 1em;
+      padding: 0.25em;
+      line-height: 48px;
+      border: 0;
+      font-size: 16px;
+      background-image: -webkit-gradient(
+        linear,
+        left top,
+        right top,
+        from(rgba(84, 147, 247, 0.9)),
+        to(rgba(84, 147, 247, 0.9))
+      );
+      color: #fff;
+      border-radius: 28px;
+      width: 100%;
+      cursor: pointer;
+
+      &:hover {
+        background-image: -webkit-gradient(
+          linear,
+          left top,
+          right top,
+          from(#228bbc),
+          to(#002c81)
+        );
+      }
+
+      &[disabled] {
+        background: rgba(84, 147, 247, 0.2);
+        cursor: default;
+
+        &:hover {
+          box-shadow: none;
+        }
+      }
+    }
+
+    i {
+      margin-right: 10px;
+    }
+  }
+
+  @media screen and (min-width: 360px) {
+    section {
+      border-radius: 5px;
+      padding: 30px 20px;
+    }
+  }
+}
