+++ frontend/src/index.scss
@@ -3,15 +3,53 @@ body {
   padding: 0;
   font-family: 'Gotham', 'Helvetica Neue', 'Arial', sans-serif !important;
   background-color: black;
-  color: #2043b5 !important;
   font-size: 16px;
   letter-spacing: -0.1px !important;
   height: 100vh;
 }
 
 a {
-  color: #2043b5;
+  color: white;
   text-decoration: none;
+}
+
+.Toastify__toast-container--top-right {
+  top: 15px !important;
+  right: 15px !important;
+  width: 360px !important;
+}
+
+.Toastify__toast {
+  word-break: break-word !important;
+  font-family: inherit !important;
+  padding: 15px !important;
+  border-radius: 5px !important;
+  font-size: 15px !important;
+  letter-spacing: 0 !important;
+  line-height: 20px !important;
+}
+
+.Toastify__toast--info {
+  background: #5493f7 !important;
+}
+
+.Toastify__toast--success {
+  background: #1daa8e !important;
+}
+
+.Toastify__toast--warning {
+  background: #fd9a02 !important;
+}
+
+.Toastify__toast--error {
+  background: #ff5561 !important;
+}
+
+.material-icons {
+  font-family: 'Material Icons';
+  font-size: 16px;
+  vertical-align: middle;
+  margin-top: -2px;
 }
 
 /* latin-ext */
